store method of classifier base class working
