+++ DB/classification/models.py
@@ -2,12 +2,7 @@ import datetime
 import pickle
 import psycopg2
 import uuid
-from sklearn.ensemble import RandomForestClassifier
 
-
-classifier = RandomForestClassifier(n_estimators=1000, max_depth=None,min_samples_split=1, max_features=3)
-print(str(classifier))
-pickle.dumps(classifier)
 
 def insert_query_builder(table=None, to_insert=None):
     attrs = list(to_insert.keys())
@@ -31,13 +26,14 @@ def post_query_builder():
 
 class ClassifierModel(object):
     """
+    :summary: Base class that all other classifiers will inherit from
     """
     model_type = None
-    date_time = None
+    created_at = None
     id = None
     model = None
 
-    def __init__(self, model_type="default", date_time=datetime.datetime.utcnow().timestamp(),
+    def __init__(self, model_type="default", created_at=datetime.datetime.utcnow(),
                  model=None):
         """
         :summary: Classirfier Model constructo
@@ -45,17 +41,17 @@ class ClassifierModel(object):
         :param model_type (optional): string, defaults to 'default',  named parameter
         """
         self.model_type = model_type
-        self.date_time = date_time
+        self.created_at = created_at
         self.model = model
         self.id = uuid.uuid4()
-        return self
 
     def store(self):
         """
         :summary: stores *trained* model to db
         """
         try:
-            blob = pickle.dumps(classifier)
+            blob = pickle.dumps(self.model)
+            self.model = blob
         except pickle.PickleError as e:
             raise ValueError("Model pickling failed", e)
 
@@ -66,14 +62,11 @@ class ClassifierModel(object):
 
         try:
             cursor = con.cursor()
-            query = "INSERT INTO classifier_model (id, date_time, model_type, model) (%(id)s, \
-                    %(date_time)s, %(model_type)s, %(model)s) RETURNING *;"
-            values = dict()
-            values['id'] = self.id
-            values['date_time'] = self.date_time
-            values['model_type'] = self.model_type
-            values['model'] = self.model
-            cursor.execute(query, values)
+            self.id = str(self.id)
+            print(self.__dict__)
+            query = insert_query_builder("classifier_model", self.__dict__)
+            print(query)
+            cursor.execute(query, self.__dict__)
         except psycopg2.Error as e:
             raise IOError("Model insert failed", e)
         finally:
@@ -83,25 +76,6 @@ class ClassifierModel(object):
     def train(self):
         raise NotImplementedError
 
-class RandomForestModel(ClassifierModel):
-    def __init__(self, model_type="RandomForest",
-                 date_time=datetime.datetime.utcnow().timestamp(),
-                 n_estimators=1000,
-                 max_depth=None,
-                 min_samples_split=1,
-                 max_features=3):
 
-        model = RandomForestClassifier(n_estimators=n_estimators,
-                                            max_depth=max_depth,
-                                            min_samples_split=min_samples_split,
-                                            max_features=max_features)
-        super(self, model_type=model_type, date_time=date_time, model=model)
-        return self
-
-
-
-
-
-
-
-
+clr = ClassifierModel()
+clr.store()
