+++ DB/puppet/modules/deploy/manifests/init.pp
@@ -24,8 +24,8 @@ CREATE TABLE data_label (
 CREATE TABLE classifier_model (
   id UUID NOT NULL,
   created_at TIMESTAMP NOT NULL,
-  type TEXT NOT NULL,
-  model TEXT NOT NULL,
+  model_type TEXT NOT NULL,
+  model TEXT NOT NULL
 );
 '
